Smoking might drop seeds.
Chance of smoking dropping a seed is based on the base seed drop rate
multiplied by the strain's seed drop mod. One problem is that currently,
when you have seeds but you run out of weed, the weed's no longer
listed, so you don't see that you have seeds. And in fact, it's probably
removed entirely, so you probably actually lose the seeds.