React Toolbox-ification, part 1.
Lots of stuff. Alerts, Buy/Sell Modal, Player Actions, and Player
Component all updated to React Toolbox style. Also main App.js and app
header bar. Some prep for Buy/Sell - the modal exists but the button to
open it is not present. Also, the player has money in their state, but
it is not displayed yet. Relates to issue #58.